CI: enable Rubocop on build
I regenerated the `.rubocop_todo.yml` file based on the current
failures, and enabled `rubocop` for all default builds. It's not really
necessary to run `rubocop` on more than one build, but for the sake of
code simplicity, I'm running it on all CI platforms and ruby versions.
If it becomes a problem that slows down builds too much, we can take
another look and figure out how best to limit it to just one build.

I added an exclude for `.spec` files. Not sure when these are used as
Ruby files, but ImageMagick compilation generates a `.spec` file that is
not Ruby, and Rubocop complains about an invalid byte sequence.